+++ hpe3parclient/__init__.py
@@ -22,7 +22,7 @@ HPE 3PAR Client.
 
 """
 
-version_tuple = (4, 2, 0)
+version_tuple = (4, 2, 1)
 
 
 def get_version_string():